feat: declare types in its own folder
